Move checkpoint utilities in a new module. Update documentation.
