+++ avalanche/benchmarks/datasets/external_datasets/fmnist.py
@@ -2,7 +2,7 @@ import dill
 from torchvision.datasets import FashionMNIST
 
 from avalanche.benchmarks.datasets import default_dataset_location
-from avalanche.training.checkpoint import constructor_based_serialization
+from avalanche.checkpointing import constructor_based_serialization
 
 
 def get_fmnist_dataset(dataset_root):
