+++ avalanche/benchmarks/datasets/lvis_dataset/lvis_dataset.py
@@ -27,7 +27,7 @@ from avalanche.benchmarks.datasets import (
     default_dataset_location,
 )
 from avalanche.benchmarks.datasets.lvis_dataset.lvis_data import lvis_archives
-from avalanche.training.checkpoint import constructor_based_serialization
+from avalanche.checkpointing import constructor_based_serialization
 
 try:
     from lvis import LVIS
