+++ avalanche/benchmarks/datasets/openloris/openloris.py
@@ -24,7 +24,7 @@ from avalanche.benchmarks.datasets import (
     default_dataset_location,
 )
 from avalanche.benchmarks.datasets.openloris import openloris_data
-from avalanche.training.checkpoint import constructor_based_serialization
+from avalanche.checkpointing import constructor_based_serialization
 
 
 class OpenLORIS(DownloadableDataset):
