+++ docs/checkpointing.rst
@@ -1,0 +1,29 @@
+Checkpointing module
+============================
+
+.. currentmodule:: avalanche.checkpointing
+
+Checkpointing allows to save and load serialized strategies to stop and resume experiments.
+
+
+Save and load utilities
+----------------------------------------
+
+.. currentmodule:: avalanche.checkpointing
+
+.. autosummary::
+    :toctree: generated
+
+    save_checkpoint
+    maybe_load_checkpoint
+
+
+Optimizing datasets for checkpointing
+----------------------------------------
+
+.. currentmodule:: avalanche.checkpointing
+
+.. autosummary::
+    :toctree: generated
+
+    constructor_based_serialization
