+++ docs/training.rst
@@ -184,21 +184,3 @@ Strategy implemented as plugins in `avalanche.training.plugins`.
     MIRPlugin
     RARPlugin
     FromScratchTrainingPlugin
-
-Utilities
-----------------------------------------
-
-Checkpointing allows to save and load serialized strategies to stop and resume experiments.
-
-Utilities
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-Utilities in `avalanche.training.checkpoint`.
-
-.. currentmodule:: avalanche.training.checkpoint
-
-.. autosummary::
-    :toctree: generated
-
-    save_checkpoint
-    maybe_load_checkpoint
